// src/app/generated/model/roleResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RoleResponse { 
    /**
     * Keycloak-internal id of the realm role.
     */
    id?: string;
    /**
     * Realm role name.
     */
    name?: string;
    /**
     * Realm role description, if one was set in Keycloak.
     */
    description?: string;
}